Deprecate old llm calling functions
